feat: проектная работа 14

+++ controllers/cards.js
@@ -1,45 +1,57 @@
+const mongoose = require('mongoose');
 const { Card } = require('../models/card');
-const { handleError } = require('../utils/error');
+const { NotFoundError } = require('../errors/NotFoundError');
+const { UnauthorizedError } = require('../errors/UnauthorizedError');
 
-async function createCard(req, res) {
+async function createCard(req, res, next) {
   try {
     const { name, link } = req.body;
     const ownerId = req.user._id;
     const card = await Card.create({ name, link, owner: ownerId });
     res.send(card);
   } catch (err) {
-    handleError(err, req, res);
+    next(err);
   }
 }
 
-async function getAllCards(req, res) {
+async function getAllCards(req, res, next) {
   try {
     const cards = await Card.find({});
     res.send(cards);
   } catch (err) {
-    handleError(err, req, res);
+    next(err);
   }
 }
 
-async function deleteCard(req, res) {
+async function deleteCard(req, res, next) {
   try {
     const { cardId } = req.params;
 
-    const card = await Card.findByIdAndRemove(cardId);
+    let card = mongoose.Types.ObjectId.isValid(cardId);
+
+    if (card) {
+      card = await Card.findById(cardId).populate('owner');
+    }
 
     if (!card) {
-      const error = new Error('Карточка не найдена');
-      error.name = 'NotFoundError';
-      throw error;
+      throw new NotFoundError('Карточка не найдена');
     }
+    const ownerId = card.owner.id;
+    const userId = req.user._id;
+
+    if (ownerId !== userId) {
+      throw new UnauthorizedError('Удалить можно только свою карточку');
+    }
+
+    await Card.findByIdAndRemove(cardId);
 
     res.send(card);
   } catch (err) {
-    handleError(err, req, res);
+    next(err);
   }
 }
 
-async function deleteLike(req, res) {
+async function deleteLike(req, res, next) {
   try {
     const userId = req.user._id;
     const card = await Card.findByIdAndUpdate(
@@ -48,17 +60,15 @@ async function deleteLike(req, res) {
       { new: true },
     );
     if (!card) {
-      const error = new Error('Карточка не найдена');
-      error.name = 'NotFoundError';
-      throw error;
+      throw new NotFoundError('Карточка не найдена');
     }
     res.send(card);
   } catch (err) {
-    handleError(err, req, res);
+    next(err);
   }
 }
 
-async function putLike(req, res) {
+async function putLike(req, res, next) {
   try {
     const userId = req.user._id;
     const card = await Card.findByIdAndUpdate(
@@ -67,13 +77,11 @@ async function putLike(req, res) {
       { new: true },
     );
     if (!card) {
-      const error = new Error('Карточка не найдена');
-      error.name = 'NotFoundError';
-      throw error;
+      throw new NotFoundError('Карточка не найдена');
     }
     res.send(card);
   } catch (err) {
-    handleError(err, req, res);
+    next(err);
   }
 }
